Basic support for custom validation messages
By default the validation messages comes from  tv4.js, but now you can
specify your own via the "validationMessages" attribute either in schema
(not recomended) or in the form defintion. If it's a string, that
is presented to the user on any error *but* required. If it's an object
that is treated as an map between tv4.js error codes and a messages.

Special "keys" are 'required' and 'default' that do what you think they do.