New list format for titleMap
titleMap can (and should) be specified as a list of name value objects, { name: ..., value: ...},
instead of just an objects with value as keys. This new format is used internally but old object
format is still accepted in form definitions.

This solves a number of problems, you can have titleMaps without enums, order of enum is
preserved when no titleMap is defined, types with titleMaps can bind agains schema types
that are not string, i.e. radiobuttons for a boolean.