Anyplace where there is a DOM field ID or NAME, or a label element FOR, change it to use a fully qualified value.  For IDs and FORs, join the key components with a '-', and for NAMEs, join key components with a '.'.
Previously, only the last component of the form.key was used to build these values (in most cases); the problem with that is when you have multiple leaves in your model with the same name for the final component.  In these cases, if you click on the label for any of these fields, focus is given to the first one instead of the current one as was probably intended.

Also, IDs are intended to be unique within the page DOM.