Fix for #121 to redraw with proper defaults
The problem was that initialForm was used to generate a form on redraw
event, but when it was an object type, it was modified in-place by
rendering function. So, when the next redraw event happened, initialForm
already had the properties defined, and it prevented the default values
from being populated into them.