Add resolution for jQuery in bower config.
Resolving as version 2.x because that's the version angular-schema-form depends on. If there's an issue with angular-ui-sortable, it the resolution should probably be increased to 3.x.

This was the original version conflict:

Unable to find a suitable version for jquery, please choose one by typing one of the numbers below:
    1) jquery#~2.1.1 which resolved to 2.1.4 and is required by angular-schema-form
    2) jquery#>=1.6 which resolved to 2.1.4 and is required by jquery-ui#1.12.1
    3) jquery#>=1.7 which resolved to 2.1.4 and is required by pickadate#3.5.6
    4) jquery#>=1.7.2 which resolved to 2.1.4 and is required by spectrum#1.7.0
    5) jquery#>=3.1.x which resolved to 3.3.1 and is required by angular-ui-sortable#0.19.0